+++ packages/@lwc/engine-core/src/framework/main.ts
@@ -79,8 +79,6 @@ export {
     setGetProperty,
     setHTMLElement,
     setInsert,
-    setInsertGlobalStylesheet,
-    setInsertStylesheet,
     setIsConnected,
     setIsHydrating,
     setIsNativeShadowDefined,
@@ -91,12 +89,11 @@ export {
     setRemove,
     setRemoveAttribute,
     setRemoveEventListener,
-    setRemoveStylesheet,
-    setRemoveGlobalStylesheet,
     setSetAttribute,
     setSetCSSStyleProperty,
     setSetProperty,
     setSetText,
     setSsr,
     setAddEventListener,
+    setToggleStyleSheet,
 } from '../renderer';
